@@ -22,25 +22,31 @@
   },
   "devDependencies": {
     "@semantic-release/commit-analyzer": "13.0.1",
-    "@semantic-release/github": "11.0.1",
-    "@semantic-release/npm": "12.0.1",
-    "@semantic-release/release-notes-generator": "14.0.3",
-    "@testing-library/jest-dom": "6.6.3",
-    "@testing-library/react": "16.1.0",
-    "@testing-library/user-event": "14.5.2",
-    "@types/lodash": "4.17.14",
+    "@semantic-release/github": "12.0.2",
+    "@semantic-release/npm": "13.1.3",
+    "@semantic-release/release-notes-generator": "14.1.0",
+    "@simonheys/eslint-config": "workspace:*",
+    "@testing-library/jest-dom": "6.9.1",
+    "@testing-library/react": "16.3.1",
+    "@testing-library/user-event": "14.6.1",
+    "@types/lodash": "4.17.21",
     "@types/lodash-es": "4.17.12",
-    "@types/react": "19.0.2",
-    "@types/react-dom": "19.0.2",
-    "@vitejs/plugin-react-swc": "3.7.2",
-    "eslint": "8.57.1",
-    "happy-dom": "16.3.0",
+    "@types/react": "19.2.7",
+    "@types/react-dom": "19.2.3",
+    "@typescript-eslint/eslint-plugin": "8.50.0",
+    "@typescript-eslint/parser": "8.50.0",
+    "@vitejs/plugin-react-swc": "4.2.2",
+    "concurrently": "9.2.1",
+    "eslint": "9.39.2",
+    "eslint-plugin-react": "7.37.5",
+    "eslint-plugin-react-hooks": "7.0.1",
+    "happy-dom": "20.0.11",
     "lodash": "4.17.21",
-    "semantic-release": "24.2.1",
-    "typescript": "5.7.2",
-    "vite": "6.0.7",
-    "vite-plugin-dts": "4.4.0",
-    "vitest": "2.1.8"
+    "semantic-release": "25.0.2",
+    "typescript": "5.9.3",
+    "vite": "7.3.0",
+    "vite-plugin-dts": "4.5.4",
+    "vitest": "4.0.16"
   },
   "peerDependencies": {
     "lodash-es": "^4.17.21",
@@ -50,7 +56,9 @@
   },
   "scripts": {
     "build": "vite build",
-    "dev": "cd example && pnpm dev",
+    "build:watch": "vite build --watch",
+    "dev": "concurrently \"pnpm build:watch\" \"pnpm dev:example\"",
+    "dev:example": "cd example && pnpm dev",
     "publish-npm": "pnpm build && npm publish --public",
     "test": "vitest run",
     "test:watch": "vitest",
