@@ -11,10 +11,10 @@
 #import "LogicParserStringUtil.h"
 // #import "NSLog.h"
 
-NSArray *LogicParserEqualityOperators;
-NSArray *LogicParserMathOperators;
-NSArray *LogicParserBooleanOperators;
-NSArray *LogicParserConversionOperators;
+extern NSArray *LogicParserEqualityOperators;
+extern NSArray *LogicParserMathOperators;
+extern NSArray *LogicParserBooleanOperators;
+extern NSArray *LogicParserConversionOperators;
 
 @interface LogicParser : NSObject {
     int _day;
