@@ -1,25 +1,22 @@
-import { CSSProperties, FC, PropsWithChildren } from "react";
+import { CSSProperties, FC, PropsWithChildren } from 'react'
 
 export interface WordClockWordProps extends PropsWithChildren {
-  highlighted: boolean;
+  highlighted: boolean
 }
 
 const wordStyle: CSSProperties = {
-  marginRight: "0.2em",
-  transition: "color 0.15s",
-};
+  marginRight: '0.2em',
+  transition: 'color 0.15s',
+}
 
-export const WordClockWord: FC<WordClockWordProps> = ({
-  highlighted,
-  ...rest
-}) => {
+export const WordClockWord: FC<WordClockWordProps> = ({ highlighted, ...rest }) => {
   return (
     <div
       style={{
         ...wordStyle,
-        color: highlighted ? "#ff0000" : "inherit",
+        color: highlighted ? '#ff0000' : 'inherit',
       }}
       {...rest}
     />
-  );
-};
+  )
+}
