@@ -1,5 +1,5 @@
-export * from "./types";
-export * from "./useWordClock";
-export * from "./WordClock";
-export * from "./WordClockContent";
-export * from "./WordClockWord";
+export * from './types'
+export * from './useWordClock'
+export * from './WordClock'
+export * from './WordClockContent'
+export * from './WordClockWord'
