@@ -6,15 +6,15 @@
 //  Copyright (c) Studio Heys Limited. All rights reserved.
 //
 
-#import "math.h"
+#include <math.h>
 
 #define SIN_TABLE_SIZE 4096
 #define SIN_TABLE_BITMASK 4095
 
-float _sinTable[SIN_TABLE_SIZE];
-float _cosTable[SIN_TABLE_SIZE];
+extern float _sinTable[SIN_TABLE_SIZE];
+extern float _cosTable[SIN_TABLE_SIZE];
 
-void buildSinAndCosTables();
+void buildSinAndCosTables(void);
 
 static inline float getSinFromTable(float d) {
     return _sinTable[(int)(d * 651.8986469f) & SIN_TABLE_BITMASK];
