@@ -1,15 +1,14 @@
-import { keychainProfile } from "./config";
-import { getDmgInfo } from "./meta";
-import { spawnCommand } from "./util";
+import { keychainProfile } from './config'
+import { getDmgInfo } from './meta'
+import { spawnCommand } from './util'
+;(async () => {
+  const dmgInfo = await getDmgInfo()
+  const { dmgPath } = dmgInfo
 
-(async () => {
-  const dmgInfo = await getDmgInfo();
-  const { dmgPath } = dmgInfo;
-
-  let result;
+  let result
 
   try {
-    console.log("Notarizing...");
+    console.log('Notarizing...')
     const altoolResult = await spawnCommand(`xcrun`, [
       `notarytool`,
       `submit`,
@@ -17,22 +16,22 @@ import { spawnCommand } from "./util";
       `--keychain-profile`,
       keychainProfile,
       `--wait`,
-    ]);
-    console.log("altoolResult", JSON.stringify(altoolResult, null, 2));
-    const { stdout } = altoolResult;
-    result = stdout;
+    ])
+    console.log('altoolResult', JSON.stringify(altoolResult, null, 2))
+    const { stdout } = altoolResult
+    result = stdout
   } catch (e: any) {
-    console.log("altoolResult e", JSON.stringify(e, null, 2));
-    if ("stderr" in e) {
-      const { stderr } = e;
-      result = stderr;
+    console.log('altoolResult e', JSON.stringify(e, null, 2))
+    if ('stderr' in e) {
+      const { stderr } = e
+      result = stderr
     }
   }
 
-  console.log("result", result);
+  console.log('result', result)
 
   if (result) {
-    await spawnCommand(`xcrun`, [`stapler`, `staple`, `-v`, dmgPath]);
-    console.log("Done");
+    await spawnCommand(`xcrun`, [`stapler`, `staple`, `-v`, dmgPath])
+    console.log('Done')
   }
-})();
+})()
